Three files that include the update side of things
U_fetchuser fetches the user data to display on the form
update is the form that will be the display
the update_handler is the sql code that takes care of the update process.

Once the profile section is built we could have a button that leads to the update.php section to start the process.